remove package registries. move api routes to http package
