+++ command/daemon.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/valist-io/valist/core/client"
 	"github.com/valist-io/valist/core/config"
-	"github.com/valist-io/valist/registry"
+	"github.com/valist-io/valist/http"
 )
 
 // Daemon runs the valist node daemon indefinitely.
@@ -33,21 +33,23 @@ func Daemon(ctx context.Context) error {
 
 `)
 
-	apiServer, err := registry.NewServer(client, config)
+	server, err := http.NewServer(client, config)
 	if err != nil {
 		return err
 	}
-	go apiServer.ListenAndServe() //nolint:errcheck
+
+	go server.ListenAndServe() //nolint:errcheck
+	fmt.Println("server running on", server.Addr)
 
 	quit := make(chan os.Signal, 1)
 	signal.Notify(quit, os.Interrupt, syscall.SIGINT, syscall.SIGTERM)
 
 	<-quit
-	fmt.Println("Shutting down")
+	fmt.Println("shutting down")
 
 	ctx, cancel := context.WithTimeout(ctx, 30*time.Second)
 	defer cancel()
 
-	apiServer.Shutdown(ctx) //nolint:errcheck
+	server.Shutdown(ctx) //nolint:errcheck
 	return nil
 }
