--- registry/docker/testdata/Dockerfile
@@ -1,2 +1,0 @@
-FROM alpine
-CMD ["echo", "Hello World!"]
