--- registry/npm/testdata/v0/.npmrc
@@ -1,1 +1,0 @@
-//localhost:10001/:_authToken=dummy
