--- registry/npm/testdata/v0/index.js
@@ -1,1 +1,0 @@
-console.log('Hello world!');
