--- registry/npm/testdata/v0/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "@valist/sdk",
-  "version": "0.0.1",
-  "description": "Valist SDK",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "license": "ISC"
-}
