--- registry/npm/testdata/v1/index.js
@@ -1,1 +1,0 @@
-console.log('Hello world!');
