--- registry/npm/testdata/v1/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "@valist/sdk",
-  "version": "0.0.2",
-  "description": "Valist SDK",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "license": "ISC"
-}
